Address PR review findings
Backend:
- Add 30s WS read timeout + ping recovery on Loki tail
- Clamp tail_lines to [1, 5000] in query_range
- Skip non-finite timestamps/counts in build_count_buckets
- Accept numeric or string counts in LokiMetricQueryResponse
- Match anyhow chain with {:#} for "exceeds the limit" -> 400 mapping
- Tests for numeric timestamp and numeric count deserialization

Frontend:
- Route Refresh through historical path for custom range / bucket
- Stop rebinding IntersectionObserver per streamed line
- Generation guard on loadHistoricalLogs
- Lazy-load recharts and react-day-picker into separate chunks
- Sort loadOlder results before concat; fold selectedBucket clear
  into change handlers
- Drive chart and log-surface colors from CSS vars (theme-aware)
- Reorder tooltip to match bar stack order; memoize highlightLogJson

Helm + docs:
- Guard bundled-Loki retention fallback against missing limits_config
- Tighten Alloy bearer-token preflight to require RISE_LOKI_BEARER_TOKEN
  in extraEnv with a descriptive error
- Scope CI helm repo config to runner.temp (no global mutation)
- Document the cluster-wide Alloy pod watch as a deliberate trade-off
- Clarify snake_case settings vs camelCase Helm values label keys
