+++ frontend/src/styles/rise.css
@@ -36,6 +36,14 @@
     --radius: 6px;
     --radius-sm: 5px;
     --radius-xs: 3px;
+
+    /* Log surface and chart stack colors. Read by the deployment logs UI
+       (JS uses var() to fill chart bars; CSS uses for row gradients). */
+    --r-logs-bg: oklch(0.985 0.004 80);
+    --r-logs-bg-hover: oklch(0.94 0.005 80);
+    --r-log-chart-info: oklch(0.55 0.005 80);
+    --r-log-chart-warn: oklch(0.70 0.14 85);
+    --r-log-chart-error: oklch(0.58 0.18 25);
 }
 
 .dark {
@@ -68,6 +76,12 @@
     --shadow-sm: 0 1px 2px rgba(0, 0, 0, 0.3);
     --shadow-md: 0 4px 18px rgba(0, 0, 0, 0.4);
     --shadow-lg: 0 24px 60px rgba(0, 0, 0, 0.5);
+
+    --r-logs-bg: oklch(0.14 0.005 80);
+    --r-logs-bg-hover: oklch(0.22 0.005 80);
+    --r-log-chart-info: oklch(0.62 0.005 80);
+    --r-log-chart-warn: oklch(0.78 0.13 85);
+    --r-log-chart-error: oklch(0.62 0.16 25);
 }
 
 /* Palette: shifts the accent plus a subtle tint of the light-mode
@@ -1091,8 +1105,8 @@ textarea.r-field { resize: vertical; min-height: 80px; font-family: 'JetBrains M
 
 /* ---------- Logs ---------- */
 .r-logs {
-    background: oklch(0.14 0.005 80);
-    color: oklch(0.92 0.005 80);
+    background: var(--r-logs-bg);
+    color: var(--text);
     font-family: 'JetBrains Mono', monospace;
     font-size: 12px;
     line-height: 1.62;
@@ -1101,7 +1115,7 @@ textarea.r-field { resize: vertical; min-height: 80px; font-family: 'JetBrains M
     height: 100%;
     border-radius: 5px;
 }
-.dark .r-logs { background: oklch(0.12 0.004 80); }
+.dark .r-logs { color: oklch(0.92 0.005 80); }
 .r-logs .ts { color: oklch(0.55 0.005 80); margin-right: 10px; }
 .r-logs .src { color: oklch(0.72 0.10 165); margin-right: 8px; }
 .r-logs .lv-info { color: oklch(0.82 0.005 80); }
@@ -1122,7 +1136,7 @@ textarea.r-field { resize: vertical; min-height: 80px; font-family: 'JetBrains M
 }
 .r-logs-row:hover:not(.is-expanded),
 .r-logs-row:focus-visible:not(.is-expanded) {
-    background: oklch(0.22 0.005 80);
+    background: var(--r-logs-bg-hover);
 }
 .r-logs-row.is-expanded {
     white-space: pre-wrap;
@@ -1136,10 +1150,10 @@ textarea.r-field { resize: vertical; min-height: 80px; font-family: 'JetBrains M
     bottom: 0;
     width: 60px;
     pointer-events: none;
-    background: linear-gradient(to right, transparent, oklch(0.14 0.005 80) 70%);
+    background: linear-gradient(to right, transparent, var(--r-logs-bg) 70%);
 }
 .r-logs-row:hover:not(.is-expanded)::after {
-    background: linear-gradient(to right, transparent, oklch(0.22 0.005 80) 70%);
+    background: linear-gradient(to right, transparent, var(--r-logs-bg-hover) 70%);
 }
 .r-logs-row .r-logs-ts {
     color: oklch(0.55 0.005 80);
